src/google_ads_excluder/requirements.txt
google-ads==18.0.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
google-api-python-client==2.55.0
google-cloud-bigquery==3.3.0
google-cloud-pubsub==2.13.4
google-cloud-storage==2.2.1
jsonschema==4.9.1
pandas==1.4.3
